+++ apps/web/src/components.d.ts
@@ -18,7 +18,6 @@ declare module 'vue' {
     DsfrBadge: typeof import('@gouvminint/vue-dsfr')['DsfrBadge']
     DsfrBreadcrumb: typeof import('@gouvminint/vue-dsfr')['DsfrBreadcrumb']
     DsfrButton: typeof import('@gouvminint/vue-dsfr')['DsfrButton']
-    DsfrCard: typeof import('@gouvminint/vue-dsfr')['DsfrCard']
     DsfrCheckbox: typeof import('@gouvminint/vue-dsfr')['DsfrCheckbox']
     DsfrCheckboxSet: typeof import('@gouvminint/vue-dsfr')['DsfrCheckboxSet']
     DsfrHighlight: typeof import('@gouvminint/vue-dsfr')['DsfrHighlight']
@@ -83,6 +82,9 @@ declare module 'vue' {
     TeeForm: typeof import('./components/element/form/TeeForm.vue')['default']
     TeeFormCallback: typeof import('./components/element/form/TeeFormCallback.vue')['default']
     TeeFormElement: typeof import('./components/element/form/TeeFormElement.vue')['default']
+    TeeFormElementCheckbox: typeof import('./components/element/form/TeeFormElementCheckbox.vue')['default']
+    TeeFormElementInput: typeof import('./components/element/form/TeeFormElementInput.vue')['default']
+    TeeFormElementSelect: typeof import('./components/element/form/TeeFormElementSelect.vue')['default']
     TeeGroupBar: typeof import('./components/element/TeeGroupBar.vue')['default']
     TeeHeader: typeof import('./components/TeeHeader.vue')['default']
     TeeHomeInfos: typeof import('./components/home/TeeHomeInfos.vue')['default']
