+++ apps/web/src/components/element/form/TeeFormElement.vue
@@ -1,104 +1,34 @@
 <template>
-  <div
-    v-if="localField.type === FieldType.Checkbox"
-    :class="`fr-col-${localField.colSize || '12'} fr-col-md-${localField.colSize || '12'}`"
-  >
-    <DsfrCheckbox
-      v-model="localField.value"
-      :name="`form-checkbox-${localField.label}`"
-      :is-valid="localField.isValid"
-      :required="localField.required"
-      @update:model-value="validateFormField"
-      @focusout="validateFormField"
-    >
-      <template #label>
-        <span> {{ localField.label }} <code>*</code></span>
-      </template>
-    </DsfrCheckbox>
-    <!-- CHECKBOX HINT -->
-    <span class="fr-hint-text fr-mt-5v">
-      {{ localField.hint }}
-      <router-link
-        v-if="localField.hintLink"
-        :to="{ name: localField.hintLink.route }"
-        target="_blank"
-      >
-        {{ localField.hintLink.text }}
-      </router-link>
-      .
-    </span>
-  </div>
-  <div
-    v-else-if="localField.type === FieldType.Select"
-    :class="`fr-col-${localField.colSize || '12'} fr-col-md-${localField.colSize || '12'}`"
-  >
-    <DsfrInputGroup
-      :error-message="getErrorMessage()"
-      :valid-message="getValidMessage()"
-    >
-      <DsfrSelect
-        v-model="localField.value"
-        :label="localField.label"
-        :required="localField.required"
-        :name="`form-select-${localField.label}`"
-        :options="localField.options"
-      />
-    </DsfrInputGroup>
-  </div>
-  <div
-    v-else
-    :class="`fr-col-${localField.colSize || '12'} fr-col-md-${localField.colSize || '12'}`"
-  >
-    <DsfrInputGroup
-      :error-message="getErrorMessage()"
-      :valid-message="getValidMessage()"
-    >
-      <DsfrInput
-        v-model="localField.value"
-        class="fr-bg--white"
-        :type="localField.type"
-        label-visible
-        :is-valid="localField.isValid"
-        :required="localField.required"
-        :label="localField.label"
-        :rows="localField.rows"
-        :is-textarea="!!localField.rows"
-        :wrapper-class="localField.wrapperClass"
-        :hint="field.hint"
-        @focusout="validateFormField"
-      >
-        <template
-          v-if="localField.callOut"
-          #label
-        >
-          {{ localField.label }}
-          <slot name="required-tip">
-            <span
-              v-if="localField.required"
-              class="required"
-              >*</span
-            >
-          </slot>
-
-          <TeeCallout
-            class="fr-bg--blue fr-text--white fr-px-2v fr-pt-2v fr-pb-0 fr-mb-0 fr-text--bold"
-            :type="localField.callOut.type"
-            :img="`${publicPath}${localField.callOut.img}`"
-            :img-container-class="'fr-col-xl-2 fr-hidden fr-unhidden-lg'"
-            :content-class="'fr-pb-2v fr-px-3v fr-px-lg-0'"
-          >
-            {{ localField.callOut.content }}
-          </TeeCallout>
-        </template>
-      </DsfrInput>
-    </DsfrInputGroup>
+  <div :class="`fr-col-${localField.colSize || '12'} fr-col-md-${localField.colSize || '12'}`">
+    <component
+      :is="currentComponent"
+      :field="localField"
+      :public-path="publicPath"
+      :get-error-message="getErrorMessage"
+      :get-valid-message="getValidMessage"
+      @update-field="(value) => validateFormField(value as DefaultFieldFormType['value'])"
+    />
   </div>
 </template>
 <script setup lang="ts">
 import { FieldType, DefaultFieldFormType, isValidatedStringFieldInputType } from '@/types'
 import Config from '@/config'
+import TeeFormElementCheckbox from './TeeFormElementCheckbox.vue'
+import TeeFormElementSelect from './TeeFormElementSelect.vue'
+import TeeFormElementInput from './TeeFormElementInput.vue'
 
 const publicPath = Config.publicPath !== 'undefined/' ? Config.publicPath : '../../public/'
+
+const currentComponent = computed(() => {
+  switch (localField.value.type) {
+    case FieldType.Checkbox:
+      return TeeFormElementCheckbox
+    case FieldType.Select:
+      return TeeFormElementSelect
+    default:
+      return TeeFormElementInput
+  }
+})
 
 interface Props {
   field: DefaultFieldFormType
@@ -112,7 +42,8 @@ const isFieldValid = (): boolean => {
   return localField.value.value !== undefined && localField.value.value !== '' && localField.value.value !== false
 }
 
-const validateFormField = (): void => {
+const validateFormField = (value: DefaultFieldFormType['value']): void => {
+  localField.value.value = value
   if (isValidatedStringFieldInputType(localField.value)) {
     localField.value.isValid = localField.value.validation(localField.value.value, !!localField.value.label?.includes('SIRET')) as boolean
   } else {
